Moves light-curve simulation logic into SN models
